refactoring drum params panel
